+++ test/flop_phoenix_test.exs
@@ -1,5 +1,6 @@
 defmodule Flop.PhoenixTest do
   use ExUnit.Case
+  use Phoenix.Component
   use Phoenix.HTML
 
   import Flop.Phoenix
@@ -31,24 +32,78 @@ defmodule Flop.PhoenixTest do
     |> Floki.parse_fragment!()
   end
 
-  defp render_table(assigns \\ []) do
-    assigns =
-      Keyword.merge(
-        [
-          __changed__: nil,
-          headers: ["name"],
-          items: [%{name: "George"}],
-          meta: %Flop.Meta{flop: %Flop{}},
-          path_helper: &route_helper/3,
-          path_helper_args: [%{}, :index],
-          row_func: fn %{name: name}, _opts -> [name] end
-        ],
-        assigns
-      )
+  defp render_table(assigns \\ [], component \\ &test_table/1) do
+    assigns = Keyword.put(assigns, :__changed__, nil)
 
-    (&table/1)
+    component
     |> render_component(assigns)
     |> Floki.parse_fragment!()
+  end
+
+  defp test_table(assigns) do
+    assigns =
+      assigns
+      |> assign_new(:for, fn -> nil end)
+      |> assign_new(:event, fn -> nil end)
+      |> assign_new(:items, fn ->
+        [%{name: "George", email: "george@george.pet", age: 8, species: "dog"}]
+      end)
+      |> assign_new(:meta, fn -> %Flop.Meta{flop: %Flop{}} end)
+      |> assign_new(:path_helper, fn -> &route_helper/3 end)
+      |> assign_new(:path_helper_args, fn -> [%{}, :index] end)
+      |> assign_new(:opts, fn -> [] end)
+      |> assign_new(:target, fn -> nil end)
+
+    ~H"""
+    <Flop.Phoenix.table
+      for={@for}
+      event={@event}
+      items={@items}
+      meta={@meta}
+      opts={@opts}
+      path_helper={@path_helper}
+      path_helper_args={@path_helper_args}
+      target={@target}
+    >
+      <:col let={pet} label="Name" field={:name}><%= pet.name %></:col>
+      <:col let={pet} label="Email" field={:email}><%= pet.email %></:col>
+      <:col let={pet} label="Age"><%= pet.age %></:col>
+      <:col let={pet} label="Species" field={:species}><%= pet.species %></:col>
+    </Flop.Phoenix.table>
+    """
+  end
+
+  defp test_table_with_foot(assigns) do
+    ~H"""
+    <Flop.Phoenix.table
+      event="sort"
+      items={[%{name: "George"}]}
+      meta={%Flop.Meta{flop: %Flop{}}}
+    >
+      <:col let={pet} label="Name" field={:name}><%= pet.name %></:col>
+      <:foot>
+        <tr><td>snap</td></tr>
+      </:foot>
+    </Flop.Phoenix.table>
+    """
+  end
+
+  defp test_table_with_html_header(assigns) do
+    ~H"""
+    <Flop.Phoenix.table
+      event="sort"
+      items={[%{name: "George"}]}
+      meta={%Flop.Meta{flop: %Flop{}}}
+    >
+      <:col
+        let={pet}
+        label={{:safe, "<span>Hello</span>"}}
+        field={:name}
+      >
+        <%= pet.name %>
+      </:col>
+    </Flop.Phoenix.table>
+    """
   end
 
   defp route_helper(%{}, path, query) do
@@ -845,9 +900,9 @@ defmodule Flop.PhoenixTest do
         )
 
       assert [_] = Floki.find(html, "tr.mungo")
-      assert [_] = Floki.find(html, "th.bean")
+      assert [_, _, _, _] = Floki.find(html, "th.bean")
       assert [_] = Floki.find(html, "tr.salt")
-      assert [_] = Floki.find(html, "td.tolerance")
+      assert [_, _, _, _] = Floki.find(html, "td.tolerance")
     end
 
     test "doesn't render table if items list is empty" do
@@ -855,40 +910,32 @@ defmodule Flop.PhoenixTest do
     end
 
     test "displays headers without sorting function" do
-      html = render_table(headers: ["Name", "Age"])
-      assert [th] = Floki.find(html, "th:fl-contains('Name')")
-      assert Floki.children(th, include_text: false) == []
+      html = render_table()
       assert [th] = Floki.find(html, "th:fl-contains('Age')")
       assert Floki.children(th, include_text: false) == []
     end
 
-    test "displays headers with safe HTML values" do
-      html = render_table(headers: [{:safe, "<span>Hello</span>"}])
-      assert [span] = Floki.find(html, "th span")
-      assert Floki.text(span) == "Hello"
-    end
-
     test "displays headers with sorting function" do
-      html = render_table(headers: ["Name", {"Age", :age}])
+      html = render_table()
 
-      assert [th] = Floki.find(html, "th:fl-contains('Name')")
-      assert Floki.children(th, include_text: false) == []
-
-      assert [a] = Floki.find(html, "th a:fl-contains('Age')")
+      assert [a] = Floki.find(html, "th a:fl-contains('Name')")
       assert Floki.attribute(a, "data-phx-link") == ["patch"]
       assert Floki.attribute(a, "data-phx-link-state") == ["push"]
 
       assert Floki.attribute(a, "href") == [
-               "/index?order_directions[]=asc&order_by[]=age"
+               "/index?order_directions[]=asc&order_by[]=name"
              ]
     end
 
-    test "adds aria-sort attribute to first ordered field" do
-      headers = [{"Name", :name}, {"Email", :email}, {"Age", :age}]
+    test "displays headers with safe HTML values" do
+      html = render_table([], &test_table_with_html_header/1)
+      assert [span] = Floki.find(html, "th a span")
+      assert Floki.text(span) == "Hello"
+    end
 
+    test "adds aria-sort attribute to first ordered field" do
       html =
         render_table(
-          headers: headers,
           meta: %Flop.Meta{
             flop: %Flop{
               order_by: [:email, :name],
@@ -897,87 +944,75 @@ defmodule Flop.PhoenixTest do
           }
         )
 
-      assert [th_name, th_email, th_age] = Floki.find(html, "th")
+      assert [th_name, th_email, th_age, th_species] = Floki.find(html, "th")
       assert Floki.attribute(th_name, "aria-sort") == []
       assert Floki.attribute(th_email, "aria-sort") == ["ascending"]
       assert Floki.attribute(th_age, "aria-sort") == []
+      assert Floki.attribute(th_species, "aria-sort") == []
 
       html =
         render_table(
-          headers: headers,
           meta: %Flop.Meta{
             flop: %Flop{
-              order_by: [:age, :email, :age],
+              order_by: [:name, :email],
               order_directions: [:desc, :asc]
             }
           }
         )
 
-      assert [th_name, th_email, th_age] = Floki.find(html, "th")
-      assert Floki.attribute(th_name, "aria-sort") == []
+      assert [th_name, th_email, th_age, th_species] = Floki.find(html, "th")
+      assert Floki.attribute(th_name, "aria-sort") == ["descending"]
       assert Floki.attribute(th_email, "aria-sort") == []
-      assert Floki.attribute(th_age, "aria-sort") == ["descending"]
+      assert Floki.attribute(th_age, "aria-sort") == []
+      assert Floki.attribute(th_species, "aria-sort") == []
 
       html =
         render_table(
-          headers: headers,
           meta: %Flop.Meta{flop: %Flop{order_by: [], order_directions: []}}
         )
 
-      assert [th_name, th_email, th_age] = Floki.find(html, "th")
+      assert [th_name, th_email, th_age, th_species] = Floki.find(html, "th")
       assert Floki.attribute(th_name, "aria-sort") == []
       assert Floki.attribute(th_email, "aria-sort") == []
       assert Floki.attribute(th_age, "aria-sort") == []
+      assert Floki.attribute(th_species, "aria-sort") == []
     end
 
     test "renders links with click handler" do
-      html = render_table(event: "sort", headers: ["Name", {"Age", :age}])
+      html = render_table(event: "sort")
 
-      assert [th] = Floki.find(html, "th:fl-contains('Name')")
-      assert Floki.children(th, include_text: false) == []
-
-      assert [a] = Floki.find(html, "th a:fl-contains('Age')")
+      assert [a] = Floki.find(html, "th a:fl-contains('Name')")
       assert Floki.attribute(a, "href") == ["#"]
       assert Floki.attribute(a, "phx-click") == ["sort"]
-      assert Floki.attribute(a, "phx-value-order") == ["age"]
+      assert Floki.attribute(a, "phx-value-order") == ["name"]
+
+      assert [a] = Floki.find(html, "th a:fl-contains('Email')")
+      assert Floki.attribute(a, "href") == ["#"]
+      assert Floki.attribute(a, "phx-click") == ["sort"]
+      assert Floki.attribute(a, "phx-value-order") == ["email"]
     end
 
     test "adds phx-target to header links" do
-      html =
-        render_table(
-          event: "sort",
-          headers: ["Name", {"Age", :age}],
-          target: "here"
-        )
+      html = render_table(event: "sort", target: "here")
 
-      assert [_] = Floki.find(html, "th:fl-contains('Name')")
-      assert [a] = Floki.find(html, "a")
+      assert [a] = Floki.find(html, "th a:fl-contains('Name')")
       assert Floki.attribute(a, "href") == ["#"]
       assert Floki.attribute(a, "phx-click") == ["sort"]
       assert Floki.attribute(a, "phx-target") == ["here"]
-      assert Floki.attribute(a, "phx-value-order") == ["age"]
+      assert Floki.attribute(a, "phx-value-order") == ["name"]
     end
 
     test "checks for sortability if for option is set" do
       # without :for option
-      html =
-        render_table(
-          headers: [{"Name", :name}, {"Age", :age}, {"Species", :species}]
-        )
+      html = render_table()
 
       assert [_] = Floki.find(html, "a:fl-contains('Name')")
-      assert [_] = Floki.find(html, "a:fl-contains('Age')")
       assert [_] = Floki.find(html, "a:fl-contains('Species')")
 
       # with :for assign
-      html =
-        render_table(
-          for: Flop.Phoenix.Pet,
-          headers: [{"Name", :name}, {"Age", :age}, {"Species", :species}]
-        )
+      html = render_table(for: Flop.Phoenix.Pet)
 
       assert [_] = Floki.find(html, "a:fl-contains('Name')")
-      assert [_] = Floki.find(html, "a:fl-contains('Age')")
       assert [] = Floki.find(html, "a:fl-contains('Species')")
     end
 
@@ -985,7 +1020,6 @@ defmodule Flop.PhoenixTest do
       html =
         render_table(
           for: Pet,
-          headers: [{"Name", :name}],
           meta:
             build(
               :meta_on_second_page,
@@ -1013,35 +1047,45 @@ defmodule Flop.PhoenixTest do
           }
         )
 
-      assert Floki.find(html, "span.order-direction") == []
+      assert Floki.find(
+               html,
+               "a:fl-contains('Email') + span.order-direction"
+             ) == []
 
       html =
         render_table(
-          headers: [{"Name", :name}],
           meta: %Flop.Meta{
-            flop: %Flop{order_by: [:name], order_directions: [:asc]}
+            flop: %Flop{order_by: [:email], order_directions: [:asc]}
           }
         )
 
-      assert [span] = Floki.find(html, "span.order-direction")
+      assert [span] =
+               Floki.find(
+                 html,
+                 "th a:fl-contains('Email') + span.order-direction"
+               )
+
       assert Floki.text(span) == "▴"
 
       html =
         render_table(
-          headers: [{"Name", :name}],
           meta: %Flop.Meta{
-            flop: %Flop{order_by: [:name], order_directions: [:desc]}
+            flop: %Flop{order_by: [:email], order_directions: [:desc]}
           }
         )
 
-      assert [span] = Floki.find(html, "span.order-direction")
+      assert [span] =
+               Floki.find(
+                 html,
+                 "th a:fl-contains('Email') + span.order-direction"
+               )
+
       assert Floki.text(span) == "▾"
     end
 
     test "allows to set symbol class" do
       html =
         render_table(
-          headers: [{"Name", :name}],
           meta: %Flop.Meta{
             flop: %Flop{order_by: [:name], order_directions: [:asc]}
           },
@@ -1054,7 +1098,6 @@ defmodule Flop.PhoenixTest do
     test "allows to override default symbols" do
       html =
         render_table(
-          headers: [{"Name", :name}],
           meta: %Flop.Meta{
             flop: %Flop{order_by: [:name], order_directions: [:asc]}
           },
@@ -1066,7 +1109,6 @@ defmodule Flop.PhoenixTest do
 
       html =
         render_table(
-          headers: [{"Name", :name}],
           meta: %Flop.Meta{
             flop: %Flop{order_by: [:name], order_directions: [:desc]}
           },
@@ -1077,24 +1119,6 @@ defmodule Flop.PhoenixTest do
       assert Floki.text(span) == "desc"
     end
 
-    test "passes additional assigns to row function and renders all items" do
-      html =
-        render_table(
-          items: [%{name: "George", age: 8}, %{name: "Barbara", age: 2}],
-          row_opts: [suffix: "-chan"],
-          row_func: fn %{age: age, name: name}, opts ->
-            assert Keyword.keys(opts) == [:suffix]
-            suffix = Keyword.fetch!(opts, :suffix)
-            [name <> suffix, age]
-          end
-        )
-
-      assert [_] = Floki.find(html, "td:fl-contains('George-chan')")
-      assert [_] = Floki.find(html, "td:fl-contains('8')")
-      assert [_] = Floki.find(html, "td:fl-contains('Barbara-chan')")
-      assert [_] = Floki.find(html, "td:fl-contains('8')")
-    end
-
     test "renders notice if item list is empty" do
       assert [{"p", [], ["No results."]}] = render_table(items: [])
     end
@@ -1102,39 +1126,24 @@ defmodule Flop.PhoenixTest do
     test "allows to set no_results_content" do
       assert render_table(
                items: [],
-               opts: [no_results_content: ~E"<div>Nothing!</div>"]
+               opts: [no_results_content: content_tag(:div, do: "Nothing!")]
              ) == [{"div", [], ["Nothing!"]}]
     end
 
-    test "renders table footer" do
-      html =
-        render_table(
-          footer: ["snip", content_tag(:span, "snap")],
-          opts: [
-            tfoot_tr_attrs: [class: "tfoot-row"],
-            tfoot_td_attrs: [class: "tfoot-td"]
-          ]
-        )
+    test "renders table foot" do
+      html = render_table([], &test_table_with_foot/1)
 
       assert [
                {"table", [{"class", "sortable-table"}],
                 [
                   {"thead", _, _},
                   {"tbody", _, _},
-                  {"tfoot", [],
-                   [
-                     {"tr", [{"class", "tfoot-row"}],
-                      [
-                        {"td", [{"class", "tfoot-td"}], ["snip"]},
-                        {"td", [{"class", "tfoot-td"}],
-                         [{"span", [], ["snap"]}]}
-                      ]}
-                   ]}
+                  {"tfoot", [], [{"tr", [], [{"td", [], ["snap"]}]}]}
                 ]}
              ] = html
     end
 
-    test "does not render table footer if option is not set" do
+    test "does not render table foot if option is not set" do
       html = render_table()
 
       assert [
@@ -1145,16 +1154,11 @@ defmodule Flop.PhoenixTest do
 
     test "does not require path_helper when passing event" do
       html =
-        (&table/1)
-        |> render_component(
-          __changed__: nil,
-          headers: [{"Name", :name}],
-          items: [%{name: "George"}],
-          meta: %Flop.Meta{flop: %Flop{}},
-          row_func: fn %{name: name}, _opts -> [name] end,
-          event: "sort-table"
+        render_table(
+          event: "sort-table",
+          path_helper: nil,
+          path_helper_args: nil
         )
-        |> Floki.parse_fragment!()
 
       assert [link] = Floki.find(html, "a:fl-contains('Name')")
 
@@ -1162,16 +1166,56 @@ defmodule Flop.PhoenixTest do
       assert Floki.attribute(link, "href") == ["#"]
     end
 
+    test "raises if no cols are passed" do
+      assert_raise RuntimeError,
+                   ~r/^You need to add at least one `<:col>`/,
+                   fn ->
+                     render_component(&table/1,
+                       __changed__: nil,
+                       event: "sort",
+                       items: [],
+                       meta: %Flop.Meta{flop: %Flop{}}
+                     )
+                   end
+    end
+
+    test "raises if no items are passed" do
+      assert_raise RuntimeError,
+                   ~r/^You need to set the `items` assign/,
+                   fn ->
+                     render_component(&table/1,
+                       __changed__: nil,
+                       col: fn _ -> nil end,
+                       event: "sort",
+                       meta: %Flop.Meta{flop: %Flop{}}
+                     )
+                   end
+    end
+
+    test "raises if no meta is passed" do
+      assert_raise RuntimeError,
+                   ~r/^You need to set the `meta` assign/,
+                   fn ->
+                     render_component(&table/1,
+                       __changed__: nil,
+                       col: fn _ -> nil end,
+                       event: "sort",
+                       items: []
+                     )
+                   end
+    end
+
     test "raises if neither path helper nor event are passed" do
-      assert_raise RuntimeError, fn ->
-        render_component(&table/1,
-          __changed__: nil,
-          headers: [{"Name", :name}],
-          items: [%{name: "George"}],
-          meta: %Flop.Meta{flop: %Flop{}},
-          row_func: fn %{name: name}, _opts -> [name] end
-        )
-      end
+      assert_raise RuntimeError,
+                   ~r/^Flop.Phoenix.table requires either the `path_helper`/,
+                   fn ->
+                     render_component(&table/1,
+                       __changed__: nil,
+                       col: fn _ -> nil end,
+                       items: [%{name: "George"}],
+                       meta: %Flop.Meta{flop: %Flop{}}
+                     )
+                   end
     end
   end